+++ src/Screens/Debug/SetGameYear.cs
@@ -8,9 +8,12 @@
 // work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 
 using System;
+using System.Globalization;
 using System.Linq;
 using CivOne.Enums;
 using CivOne.Graphics;
+using CivOne.Services;
+using CivOne.Services.Screen;
 using CivOne.Tasks;
 
 namespace CivOne.Screens.Debug
@@ -38,14 +41,15 @@ namespace CivOne.Screens.Debug
 			_input.Y = 97 + oy;
 		}
 
-		public string Value { get; private set; }
+		public string? Value { get; private set; }
 
-		public event EventHandler Accept, Cancel;
+		public event EventHandler? Accept, Cancel;
 
-		private void GameYear_Accept(object sender, EventArgs args)
+		private void GameYear_Accept(object? sender, EventArgs __)
 		{
-			Value = (sender as Input).Text;
-			
+			if (sender is not Input input) return;
+			Value = input.Text;
+
 			int gameYear;
 			if (!int.TryParse(Value, out gameYear) || gameYear < -4000 || gameYear > 6000)
 			{
@@ -53,23 +57,19 @@ namespace CivOne.Screens.Debug
 			}
 			else
 			{
-				Game.GameTurn = Common.YearToTurn(gameYear);
+				Game.GameTurn = _gameCalendarService.YearToTurn(gameYear);
 				GameTask.Enqueue(Message.General(TranslateFormatted("Game year set to {0}.", Game.GameYear)));
 			}
 
-			if (Accept != null)
-				Accept(this, null);
-			if (sender is Input)
-				((Input)sender)?.Close();
+			Accept?.Invoke(this, EventArgs.Empty);
+			input.Close();
 			Destroy();
 		}
 
-		private void GameYear_Cancel(object sender, EventArgs args)
+		private void GameYear_Cancel(object? sender, EventArgs __)
 		{
-			if (Cancel != null)
-				Cancel(this, null);
-			if (sender is Input)
-				((Input)sender)?.Close();
+			Cancel?.Invoke(this, EventArgs.Empty);
+			if (sender is Input input) input.Close();
 			Destroy();
 		}
 
@@ -80,17 +80,27 @@ namespace CivOne.Screens.Debug
 				DrawDialog();
 			}
 
-			if (!Common.HasScreenType<Input>())
+			if (!_screenQueryService.HasScreenType<Input>())
 			{
-				Common.AddScreen(_input);
+				_screenCommandService.AddScreen(_input);
 			}
 			return false;
 		}
 
+		private readonly IScreenQueryService _screenQueryService;
+		private readonly IScreenCommandService _screenCommandService;
+		private readonly IGameCalendarService _gameCalendarService;
+		
 		public SetGameYear()
 		{
+			_screenQueryService = ScreenServiceFactory.CreateQueryService();
+			_screenCommandService = ScreenServiceFactory.CreateCommandService();
+			_gameCalendarService = GameCalendarServiceFactory.GetCurrent();
+
+			var yearString = _gameCalendarService.TurnToYear(Game.GameTurn).ToString(CultureInfo.InvariantCulture);
 			Palette = Common.Screens.Last().OriginalColours;
-			_input = new Input(Palette, Common.TurnToYear(Game.GameTurn).ToString(), 0, 5, 11, 90 + OffsetX, 97 + OffsetY, 101, 10, 5);
+
+			_input = new Input(Palette, yearString, 0, 5, 11, 90 + OffsetX, 97 + OffsetY, 101, 10, 5);
 			_input.Accept += GameYear_Accept;
 			_input.Cancel += GameYear_Cancel;