+++ src/Screens/Input.cs
@@ -18,14 +18,14 @@ namespace CivOne.Screens
 	[ScreenResizeable]
 	internal class Input : BaseScreen
 	{
-		public event EventHandler Accept;
-		public event EventHandler Cancel;
+		public event EventHandler? Accept;
+		public event EventHandler? Cancel;
 		
 		private string _text;
-		private int _fontId;
-		private byte _textColour, _cursorColour;
-		private int _width, _height;
-		private int _maxLength;
+		private readonly int _fontId;
+		private readonly byte _textColour, _cursorColour;
+		private readonly int _height;
+		private readonly int _maxLength;
 		private int _cursorPosition = -1;
 		
 		public bool CenterTo320Coordinates { get; set; }
@@ -47,7 +47,7 @@ namespace CivOne.Screens
 				_cursorPosition = 0;
 			}
 			
-			StringBuilder sb = new StringBuilder(_text);
+			StringBuilder sb = new(_text);
 			if (_text.Length > _cursorPosition)
 			{
 				sb[_cursorPosition] = character;
@@ -121,26 +121,24 @@ namespace CivOne.Screens
 					else _cursorPosition = _text.Length;
 					return true;
 				case Key.Escape:
-					if (Cancel != null)
-						Cancel(this, null);
+					Cancel?.Invoke(this, EventArgs.Empty);
 					break;
 				case Key.Enter:
-					if (Accept != null)
-						Accept(this, null);
+					Accept?.Invoke(this, EventArgs.Empty);
 					break;
 				case Key.Delete:
-					if (_cursorPosition < _text.Length)
+					if (_cursorPosition >= _text.Length)
 					{
-						sb = new StringBuilder();
-						for (int i = 0; i < _text.Length; i++)
-						{
-							if (i == _cursorPosition) continue;
-							sb.Append(_text[i]);
-						}
-						_text = sb.ToString();
-						return true;
+						return false;
 					}
-					return false;
+					sb = new StringBuilder();
+					for (int i = 0; i < _text.Length; i++)
+					{
+						if (i == _cursorPosition) continue;
+						sb.Append(_text[i]);
+					}
+					_text = sb.ToString();
+					return true;
 				case Key.Home:
 					_cursorPosition = 0;
 					return true;
@@ -159,9 +157,9 @@ namespace CivOne.Screens
 					return AppendCharacter((char)'/');
 				default:
 					char c = args.KeyChar;
-					if (!args.Shift) c = Char.ToLower(c);
+					if (!args.Shift) c = char.ToLowerInvariant(c);
 					if (args.Key == Key.Minus) c = '-';
-					if (args.Shift && (c >= '0' && c <= '9'))
+					if (args.Shift && c >= '0' && c <= '9')
 					{
 						switch (c)
 						{
@@ -183,12 +181,12 @@ namespace CivOne.Screens
 			Destroy();
 		}
 
-		private void Resize(object sender, ResizeEventArgs args)
+		private void Resize(object? _, ResizeEventArgs __)
 		{
 			HasUpdate(0);
 		}
 		
-		public Input(Palette palette, string text, int fontId, byte textColour, byte cursorColour, int x, int y, int width, int height, int maxLength)
+		public Input(Palette palette, string text, int fontId, byte textColour, byte cursorColour, int x, int y, int _, int height, int maxLength)
 		{
 			OnResize += Resize;
 
@@ -199,12 +197,11 @@ namespace CivOne.Screens
 			_cursorColour = cursorColour;
 			X = x;
 			Y = y;
-			_width = width;
 			_height = height;
 			_maxLength = maxLength;
 		}
 		
-		public Input(Palette palette, int fontId, byte textColour, byte cursorColour, int x, int y, int width, int height, int maxLength)
+		public Input(Palette palette, int fontId, byte textColour, byte cursorColour, int x, int y, int _, int height, int maxLength)
 		{
 			OnResize += Resize;
 
@@ -215,7 +212,6 @@ namespace CivOne.Screens
 			_cursorColour = cursorColour;
 			X = x;
 			Y = y;
-			_width = width;
 			_height = height;
 			_maxLength = maxLength;
 		}