+++ src/Units/TribalHuts/Handler/BarbariansEventHandler.cs
@@ -1,5 +1,6 @@
 using CivOne.Enums;
 using CivOne.Services;
+using CivOne.Services.Random;
 using CivOne.Tiles;
 
 namespace CivOne.Units.TribalHuts
@@ -10,11 +11,11 @@ namespace CivOne.Units.TribalHuts
 		private readonly int y;
 		private readonly Map map;
 
-		private readonly Random random;
+		private readonly IRandomService random;
 
 		private readonly IGame gameInstance;
 
-		public BarbariansEventHandler(int x, int y, IGame gameInstance, Map map, Random random)
+		public BarbariansEventHandler(int x, int y, IGame gameInstance, Map map, IRandomService random)
 		{
 			this.random = random;
 			this.x = x;
@@ -44,7 +45,7 @@ namespace CivOne.Units.TribalHuts
 
 					if (hasCity || hasUnits || isOcean || isPolar) continue;
 
-					if (random.Next(0, 10) < 6) continue;
+					if (random.NextInt(0, 10) < 6) continue;
 
 					gameInstance.CreateUnit(GetRandomBarbarianUnitType(), tile.X, tile.Y, 0, true);
 
@@ -57,7 +58,7 @@ namespace CivOne.Units.TribalHuts
 
 		private UnitType GetRandomBarbarianUnitType()
 		{
-			return random.Next(0, 100) < 50 ? UnitType.Cavalry : UnitType.Legion;
+			return random.NextInt(0, 100) < 50 ? UnitType.Cavalry : UnitType.Legion;
 		}
 
 		public void PostExecute()